+++ http4k-contract/src/main/kotlin/org/http4k/contract/openapi/v3/OpenApi3.kt
@@ -35,6 +35,8 @@ import org.http4k.format.AutoMarshallingJson
 import org.http4k.format.Json
 import org.http4k.format.JsonType
 import org.http4k.lens.Header.CONTENT_TYPE
+import org.http4k.lens.ParamMeta
+import org.http4k.lens.ParamMeta.ArrayParam
 import org.http4k.lens.ParamMeta.FileParam
 import org.http4k.lens.ParamMeta.ObjectParam
 import org.http4k.lens.ParamMeta.StringParam
@@ -55,7 +57,11 @@ class OpenApi3<NODE : Any>(
 ) : ContractRenderer, ErrorResponseRenderer by errorResponseRenderer {
     private data class PathAndMethod<NODE>(val path: String, val method: Method, val pathSpec: ApiPath<NODE>)
 
-    constructor(apiInfo: ApiInfo, json: AutoMarshallingJson<NODE>, extensions: List<OpenApiExtension> = emptyList()) : this(apiInfo, json, extensions, ApiRenderer.Auto(json))
+    constructor(
+        apiInfo: ApiInfo,
+        json: AutoMarshallingJson<NODE>,
+        extensions: List<OpenApiExtension> = emptyList()
+    ) : this(apiInfo, json, extensions, ApiRenderer.Auto(json))
 
     override fun description(contractRoot: PathSegments, security: Security?, routes: List<ContractRoute>): Response {
         val allSecurities = routes.map { it.meta.security } + listOfNotNull(security)
@@ -71,7 +77,10 @@ class OpenApi3<NODE : Any>(
                         it.value.map { pam -> pam.method.name.toLowerCase() to pam.pathSpec }.toMap().toSortedMap()
                     }
                     .toSortedMap(),
-                Components(json.obj(paths.flatMap { it.pathSpec.definitions() }), json(allSecurities.filterNotNull().combineFull()))
+                Components(
+                    json.obj(paths.flatMap { it.pathSpec.definitions() }),
+                    json(allSecurities.filterNotNull().combineFull())
+                )
             )
         )
 
@@ -118,8 +127,10 @@ class OpenApi3<NODE : Any>(
         .groupBy { it.message.status.code.toString() }
         .map {
             it.key to
-                ResponseContents<NODE>(it.value
-                    .map { it.description }.toSortedSet().joinToString(","), it.value.collectSchemas())
+                ResponseContents<NODE>(
+                    it.value
+                        .map { it.description }.toSortedSet().joinToString(","), it.value.collectSchemas()
+                )
         }.toMap()
 
     private fun List<HttpMessageMeta<Response>>.collectSchemas() = groupBy { CONTENT_TYPE(it.message) }
@@ -134,13 +145,20 @@ class OpenApi3<NODE : Any>(
         .toMap()
 
     private fun ContractRoute.asOpenApiParameters() = nonBodyParams.map {
-        when (it.paramMeta) {
+        val paramMeta: ParamMeta = it.paramMeta
+        when (paramMeta) {
             ObjectParam -> SchemaParameter(it, "{}".toSchema())
             FileParam -> PrimitiveParameter(it, json {
                 obj("type" to string(FileParam.value), "format" to string("binary"))
             })
+            is ArrayParam -> PrimitiveParameter(it, json {
+                obj(
+                    "type" to string("array"),
+                    "items" to obj("type" to string(paramMeta.itemType().value))
+                )
+            })
             else -> PrimitiveParameter(it, json {
-                obj("type" to string(it.paramMeta.value))
+                obj("type" to string(paramMeta.value))
             })
         }
     }
@@ -176,7 +194,8 @@ class OpenApi3<NODE : Any>(
     private fun HttpMessageMeta<HttpMessage>.toSchemaContent(): BodyContent {
         fun exampleSchemaIsValid(schema: JsonSchema<NODE>) =
             if (example is Array<*>
-                || example is Iterable<*>) !json.fields(schema.node).toMap().containsKey("\$ref")
+                || example is Iterable<*>
+            ) !json.fields(schema.node).toMap().containsKey("\$ref")
             else apiRenderer.toSchema(object {}) != schema
 
         val jsonSchema = example